@@ -19,7 +19,7 @@ namespace QReduction.QReduction.Infrastructure.DbMappings.Domain.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [EnableCors("QReductionPolicy")]
+    //[EnableCors("QReductionPolicy")]
     [ValidateModelFilter]
     [ApiExplorerSettings(GroupName = "Admin")]
 
