@@ -18,7 +18,7 @@ namespace QReduction.Apis.Controllers.Acl
 {
     [Route("api/[controller]")]
     [ApiController]
-    [EnableCors("QReductionPolicy")]
+   // [EnableCors("QReductionPolicy")]
     [ValidateModelFilter]
     public class RoleController : CustomBaseController
     {
