@@ -17,7 +17,7 @@ namespace QReduction.QReduction.Infrastructure.DbMappings.Domain.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [EnableCors("QReductionPolicy")]
+    //[EnableCors("QReductionPolicy")]
     [ValidateModelFilter]
     public class OrganizationController : CustomBaseController
     {
