@@ -24,7 +24,7 @@ namespace QReduction.QReduction.Infrastructure.DbMappings.Domain.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [EnableCors("QReductionPolicy")]
+    //[EnableCors("QReductionPolicy")]
     [ValidateModelFilter]
     public class ShiftQueueController : CustomBaseController
     {
@@ -126,7 +126,7 @@ namespace QReduction.QReduction.Infrastructure.DbMappings.Domain.Controllers
             catch (Exception ex)
             {
 
-                throw;
+                throw ex;
             }
 
 
